add items with receipt

+++ Shelfe/src/com/fydp/shelfe/RecieptScanner.java
@@ -7,6 +7,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.Date;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import com.googlecode.tesseract.android.TessBaseAPI;
 
@@ -26,6 +28,7 @@ import android.os.Environment;
 import android.provider.MediaStore;
 import android.provider.MediaStore.Images.Media;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentTransaction;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -175,9 +178,39 @@ public class RecieptScanner extends Fragment
     		/*TextView result = (TextView) this.getActivity().findViewById(R.id.textResult);
     		result.setText(recognizedText);
     		this.getActivity().setContentView(R.layout.take_photo);*/
-    		Log.i( "ReadText", "text: " + recognizedText );
+    		
+    		
+    		itemSearch(recognizedText);
     }
 
+    public void itemSearch(String text){
+    
+    	if (text.contains("SUBTOTAL")){
+    		String items = text.split("SUBTOTAL")[0];
+    		System.out.println("text: " + items );
+	    	Pattern p = Pattern.compile("(.+)([^$]\\d+(\\.|\\‘|\\,)\\d{2})");
+	    	Matcher mr = p.matcher(items);
+	    	while (mr.find()) {
+		        Fragment itemFrag = new TempItemAdd(); 
+		        Bundle bundle = new Bundle();
+		        bundle.putString("price", mr.group(2) );
+		        bundle.putString("name", mr.group(1) );
+		        itemFrag.setArguments(bundle);
+		        // consider using Java coding conventions (upper first char class names!!!)
+		        FragmentTransaction transaction = getFragmentManager().beginTransaction();
+		        // Replace whatever is in the fragment_container view with this fragment,
+		        // and add the transaction to the back stack
+		        transaction.replace(R.id.container, itemFrag);
+		        transaction.addToBackStack(null);
+
+		        // Commit the transaction
+		        transaction.commit(); 
+	    	}
+	    	
+    	}
+    }
+    
+    
     protected void onRestoreInstanceState( Bundle savedInstanceState) throws IOException{
     	Log.i( "MakeMachine", "onRestoreInstanceState()");
     	if( savedInstanceState.getBoolean( RecieptScanner.PHOTO_TAKEN ) ) {
@@ -198,6 +231,7 @@ public class RecieptScanner extends Fragment
         } catch (IOException e) {
             Log.e("tag", "Failed to get asset file list.", e);
         }
+
         for(String filename : files) {
    
 	            InputStream in = null;
@@ -205,8 +239,10 @@ public class RecieptScanner extends Fragment
 	            try {
 	              in = assetManager.open(root + "/" + filename);
 	              File outFile = new File(Environment.getExternalStorageDirectory(), root + "/" + filename);
-	              out = new FileOutputStream(outFile);
-	              copyFile(in, out);
+	              if(!outFile.exists()){
+		              out = new FileOutputStream(outFile);
+		              copyFile(in, out);
+	             }
 	            } catch(IOException e) {
 	                Log.e("tag", "Failed to copy asset file: " + filename, e);
 	            }
